src/agents/requirements.txt
google-adk>=1.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
jira>=3.8.0
